Moon_Dance/Moon_Dance/backend/requirements.txt
numpy>=1.21.0
pandas
tk
requests
fastapi
uvicorn
pydantic
Flask
Flask-Cors
PyJWT
pyOpenSSL
celery
redis
